Extract common tests into a common spec to save repetition
